Fix JSON parsing of special chars
Bug found via Fredrik: parsing failure when a JSON value ends in `\\"`. Currently it is trying to find the complete value by looking for a `"` that isn't preceded by a `\`, but this fails when the slash is part of a previous special char like `\\`.

Changed to parse JSON values as it goes instead of trying to find the end token and then unescaping the complete value. This allow it to correctly find special chars as it goes, and has a side effect of speeding up JSON parsing by about 4x in my tests. Although this is somewhat counter-intuitive, it looks like creating a string builder as it goes is cheaper than scanning once to find the end, creating a string, and then going back to look for special chars.

Minor improvements to unescape() method to avoid creating temporary array, which may also speed up direct parsing of values in other places.