Replace teamUser and host use with accounts, add /account
Remove teamUser and hosts with teamId from CDS config. These overlap with accounts, and now that we have accounts we should use them instead. After this change, all team to account and individual team to host configuration is done via accounts as per the Contest API.

After this change:
- If you want to associate a team to IP/host or account, it must be done via an account in a contest. No more global config except for <host> mapping all team to host patterns.
- If you want a user to be able to log into the CDS, you configure them in (copy username and password to) users.xml.
- If you want auto-login, the user's password must be in accounts.json and there must be an IP or host in accounts.json or <host> in CDS config.
- User accounts are 'global', so if I can login to one contest and the same username exists in another contest that's the same account (though possibly with different a team id).
- Public users can't see anything in /accounts, trusted users can see /accounts but w/o passwords, teams can only see their own account w/o password.

Also implemented /account, which just tries to find a match in /accounts from the same contest. For now, additional CDS accounts are not added to /accounts.