Don't print balloons when problem UUIDs match
When two or more problems have matching UUIDs we should only print balloons for the first one solved. I've added a preference to disable this behaviour.

I also noticed that if a team submits multiple solutions to a problem we print multiple balloons. Likewise we should only give a balloon for the first solution to a problem (with no preference to change this).