Per-account start-status control
Adds the ability to configure patch access for accounts to change individual contest objects, and a web UI to change the start-status when you have access.

On the server side, you can add CDS config to allow an account access to patch, e.g.:
  <auth username="x" contestId="test" type="start-status" id="Sysops"/>
Individual attributes can be left out to specify "all", e.g. leave out contestId to allow patching on any contest. This is not documented until we get a little feedback and use at NAC.

/access is updated to include a "type/id" capability when you have access to patch something.

On the client side, start status items is now visible on the Details page. If the user has access (via the "start-status/<id>" capability) then the status is replaced by three buttons, just like the existing admin controls.

Also changed the Theme icon to fa-adjust.