Fix admin pause time
Changes the CDS admin page to use the same formatting of contest time as the rest of the UI (including countdown presentations) so that the time always matches, especially at pauses. This was noticed by Tobi at SWERC where it looked like the admin and live view were 1s different.