Simplified contest objects & new event feed
This was meant to be two (or three) separate PRs, but they overlap a bit and became hard to separate.

1. First, I've always been 'architecturally annoyed' by the fact that I put two methods on every contest object, which typically have very similar content: getPropertiesImpl(Map<String, Object>) to get a generic list of properties, and writeBody(JSONEncoder) to encode those properties to JSON. The latter also pulls in a dependency on the output encoding, which is kind of ugly.

These couldn't easily be merged because the first method creates a map of all properties and the second directly serializes properties. I tried implementing writeBody() generically by calling getPropertiesImpl() and then serializing them, but the performance is poor because you need to create or clear a Map each time. Instead, I came up with a Properties interface and getProperties(Properties), which allows you to store properties in a Map if necessary, or serialize properties as they are added.

I had to tweak this a bit just to maintain the same performance, but then I found a couple optimizations like skipping JSON encoding for ids. For a contest feed this improves write performance on my machine from about 220ms for a 15.4Mb contest feed before this change down to under 200ms.

2. NDJSONFeedWriter has been updated to support the new contest feed notification format. The prefix (the start for token/id) is moved out of the method so that the writer doesn't need to know the contest.

I've left the old feed support in for now, enabled by setting an ICPC_OLD_FEED system property to true.

3. Minor cleanup: remove old unused classes like IContentType and ModelWriter, and removed the getProperty(String) method.