Access to event feed cache
At past contests we've had cases where we needed to look at the event feed/cache coming from the CCS, and to do that we had to log into the CDS and find the cache file. This change adds an admin-only endpoint <contest>/log so you can access the cache from anywhere. It isn't exposed through the UI for now, will try it out at NAC before deciding how to expose (likely on the admin page).