Switch admin countdown to Contest API for GET
The admin countdown control was getting the current contest time and controlling it via a custom (non-spec) endpoint. I'm going to leave the PUT as-is for now (I don't feel like tackling calculating pause and resume times and add/remove on the client-side yet) except for removing one unused 'absolute' command. On the GET side I'm switching it over to use the Contest API and common method for formatting time.

While doing this I realized that the pause/resume was jumping a bit, so I correctly floor-ed it. Then I noticed it didn't quite match the presentations, so I tracked it down and found that the presentations weren't correctly floor-ing time either, so I fixed that and refactored the method a bit.

Also found an old web socket class that could be removed.