Clean up navigation access
Based on a comment from Nicky, I updated the web page access to match what's visible in the left navigation. I changed it so staff has access to everything but the video and admin pages, and uses the admin versions of the jsps (which mostly just have API links or full data compared to the public ones).

Also found an unused time js and jsp, deleting these.