Remove --fps and add Ctrl-D
Remove the --fps command line arg and add a Ctrl-D to toggle debug mode instead, which shows fps, resolution, and presentation plan.

Why this change? Slightly simplifies the command line, provides more info, and (most importantly) can be turned on and off dynamically.

The presentation admin can already see all of this info, so I haven't added any ability to see the client debug state or control it from the admin. When an admin is being used I'm assuming we'll either continue to get the information there, or if we're directly at a machine just toggle debug mode on and off again. If that's not the case we could add this.