Simplify keyboard/mouse binding
Prior to this change you could add a listener to any presentation to register for key or mouse events. However, this was overly complicated and only one presentation (the team client display) actually added a listener - to itself. I'm simplifying this pattern to (only) allow presentations to easily listen to their own key & mouse events.

While I'm at it I added paging support to the scrolling scoreboard presentations. Using keys 1-9 and 0 (for 10) you can select which page of the scoreboard to scroll to. This allows a human user to control the scoreboard, e.g. for streaming while discussing the contest. This is just a teaser; we'll likely want to add more controls to the scoreboard (including a 'reset' option) or other presentations in the future.