Contest API provider and endpoints
Finished /api implementation and added endpoints to /access. With this change the CDS is now compliant with the current Contest API spec and will give correct results based on the user's role, with one major caveat: endpoints are calculated by building a list of objects and properties as they are added to the contest for each role. This means that /access is correct by the end of a contest, but during a contest endpoints are missing if we haven't seen them yet (e.g. problems or submissions before the start of a contest) or individual properties might be missing if something hasn't happened yet (e.g. clarifications won't have 'to_team_id' if nobody has replied directly to a team yet).

This is still worth releasing since it as good as we can do until there are CCSs that provide an /access endpoint, and the core mechanism delivered here and AccessService won't need to change. To fix this when connected to a CCS we'll have to call its /access endpoint on startup and inject this info into the properties based on role. When reading from disk we'll have to pre-read the full contest. Once this is done we may not need to check the properties of each individual object added to the contest, but I expect that's still useful to leave in for some backwards compatibility.

/api - Added provider name and logo
/api/<contest>/access - Added endpoints and properties