Revert streams to http
Changes I made in the last year to simplify auth have inadvertently moved video streams to https. We have no reason to believe this would be a problem... but also no proof that it isn't a lot of overhead like the last time we tried this >5 years ago, nor any good way to test before Dhaka. As per decision on sysops call I am reverting streams, hopefully not at the expense of anything else.