Reduce award access
As discussed on the last sysops call, awards should not be publicly visible in general. The one exception is FTS awards before the freeze.

This change fixes it with one obvious exception: if a submission happens before the freeze and the award is given after the freeze, it won't be publicly accessible. I don't think this is worth fixing for now, and can be deferred until we update the access policy:
- it's difficult and potentially ugly (you'd need to find the problem based on award id, then look for team's submissions for that problem, and if any submission was before the freeze and solved...)
- we doubt anyone is using these awards anyway