Dhaka CMS updates
Updates to import config for Dhaka:
- React to CMS institutions.tsv column changes.
- Added reading from Staff.tab to get a better list of people, since the other methods to get this have been removed from CMS.
- Start auto-team numbering from 501 so that it is more obvious.
- Cleaned up (inaccurately named) JsonToTSVConverter to document the command line better.
- Updated contest model's version of TSVImporter to be as consistent with CMS version as possible, and bring their source slightly closer together.

CMS web services have been gone for more than 2 years and it seems safe to say they're not coming back, so removing code that was able to download and parse those files. CMSDownloadhelper and CMSLogin removed, as well as unused methods and some code in JsonToTsvConverter.