Coach view reaction support
The coach view has had some of the initial support to watch reaction videos for a couple years, but I never finished the support. This completes it with the ability to click on any submission in the details view that has a reaction and it will play. The usability is not obvious, but that's ok b/c I don't want to advertise this support until we've had a chance to test on a real-ish contest, possibly after Dhaka.
Half of this change is also renaming some of the old variable names to 'webcam' and 'desktop' to match the spec and adding a 'mode' and clearMediaPlayer() to clean up the code paths.