Avoid download collision
When downloading files from a source, there can be collisions (race condition) if there are multiple processes (CDSs) writing to the same folder. If one CDS is slightly faster, then the next one may delete and redownload the file (leaving no file for a second) or two could write to the same file at the same time. The final result is fine on disk, but either CDS could have periods where the file they just wrote doesn't exist for a second.

Solves this issue by downloading to a temp file and renaming. If the file exists by the time you're done with the correct timestamp it avoids replacing it.