Fix presentation window NPE
There's an NPE that we've hit in Dhaka a few times b/c both the main painting thread and the thumbnails can update the presentation state (current presentation plan) at the same time, and possibly setting it null while the other thread is still using it. Avoids this by using a temporary variable.
Includes an unrelated change to update the window title with each presentation update.