Avoid downstream effect of "Null file ref"
Avoid adding null entries to file references list if the file metadata can't be found. Try twice before doing so. This is a bit of a hack - we know there's a file on disk, but we can't get any details about it. Trying twice will tell us if it's a temporary/timing problem or not, and better to never add null file ref (i.e. miss telling clients about one reaction vs crashing).
Also two fixes to reaction video recorder:
- Fix potential problem with file extension length if we use another video container.
- Don't delete 0-byte downloads. This will be more annoying to clients, but avoids adding the file on another CDS and then deleting the underlying file.