Move video streams to HTTPS
We initially tried HTTPS >10 years ago, and removed it for video when we hit a performance problem that was triggered when live accidentally did a DoS attack. Since then there has been a hesitation to try it again due to lack of a systems test and worry that some clients (e.g. VLC at the time) couldn't handle HTTPS well.

All clients that I've tried now support HTTPS (it's pretty much standard now) and David did some performance tests in Dhaka which confirmed that the overhead for HTTPS is negligible. This change:
- Moves all CDS content to HTTPS (was just missing video streams before).
- Removes the block on HTTPS in video servlet.
- Renames 'trusted' to 'analyst' to match the account type.
- Moved hardcoded 'http' to be part of '<host>' so that the transport, host, and port are all calculated in one place, HTTPHelper (and much simpler anyway).