Improve UIDs for CDS clients
Coach view was using the same UID (hash of user name) for all clients trying to connect to the CDS. This resulted in a lot of repeated "Disconnecting client with existing uid" messages in the CDS log, and eventually even a CPU spike due to garbage collection threads (no real clue why, but stopping coach view fixed it).

Switched to always use the hashcode of "username + network address" for all clients and improved the error message a little.
While using the admin to test I noticed a case where thumbnail images aren't disposed of properly, fixed this as well.