Expected to Advance awards
Adds support for a new type of medal, 'expected to advance'. It is applied to the top teams with at most one per organization. Support added to command line and award utility UI (via templates).

While I was here, I decided it was past time to remove the last (?) of the 'world finals' award defaults, both because they don't apply outside of finals, and because we haven't even followed those rules at finals for years. I did the following:
- Removed setting of b from CDS admin. (we've been using templates, which don't match that path anyway)
- Removed the createWorldFinalsAwards() methods from the award util and slightly simplified what the defaults are.
- Changed the default citation for winners from 'World Champion' to just 'Champion'.