Support contests longer than a month
Who would do that, right? At least that's what I thought 15 years ago when I picked the Java int type. The max ms in int is about 29 days. (In my defence, it might have even been int minutes that long ago, i.e. 60 months) But GNY decided to leave a test contest running for over a month this year to give teams plenty of time to submit. This makes the contest go into negative time and none of the tools work.

The change is really simple... but completely pervasive. If contests can be longs, then so can freeze times, standings, last submission times, etc. May as well be able to pause a contest a month in advance too. Expect a lot of change, and odds are high I'll miss something in the first commit. :)