Team label support
We're adding team labels to the Contest API, so we need to match this throughout the tools.

This PR covers:
- The actual API change.
- Default label to the id if it isn't present (this is a bit of a hack done when the name is set - couldn't think of a cleaner way). This ensures that clients can always rely on the label and it is there even if the CCS/contest package doesn't have one.
- Impact to all downstream clients: balloon printer, coach view, presentations, award util, and CDS web views.

Could I have missed something? Sure. I did a pretty thorough search which turned up all the impacts that I could think of, but until we test at NAC we won't know for sure. For now I switched all 'user output/messages' to labels, but left the few internal log messages as id because that's what you'll need for things like API lookup.