Initial HLS support
This is the initial batch of changes to support HLS. It makes the structural/URL-facing changes required to support file-serving video formats (vs streaming), and sets the default video format to HLS (even though it's not really working yet!).

There is a *lot* of change in order to support HLS so I need to start pushing a branch, but there are numerous limitations (and likely some I still don't know about) listed in issue #812 and at the top of the HLSHandler. The current code works for one stream at a time, and hasn't been tested except through Safari.