Add null checks to JSON parser
When I wrote the code I figured skipping the null checks was fine; the code is more concise and the next line of parsing fails and returns the same result. While debugging it is a pain though, since e.g. you can't tell the debugger to stop at NPEs b/c it'll hit lots of them here while parsing a bad feed. So this is no change, but a little clearer and avoids this problem.