Image gen improvements, person photos, and missing data report
I just managed to delete my commit, so no more complete and fancy message... This refactors a lot of the image gen code, which fixes a few problems it was having since the CMS moved to (mostly) generate a contest package, and makes it mostly good enough to handle person photos as well. I've also included a 'missing data report' tsv that's generated into the contest package, and sent to Lidia to help for Sharm.