Remove default contest info
When we fixed an issue with default contest state I thought I'd be clever
and also fix the default contest info. This was a bad idea, because a
default contest info has no id or name, since we don't know that yet.
And having a contest info with the wrong id, etc. in the history causes
worse problem, like not merging correctly when there is new data,
causing the / endpoint to return two contests, scoreboards to fail, ...

There was no need to make this part of the change, so let's remove it. :)