Read problem time limit from problem.yaml
There's a new (proposed) standard for specifying problem time limits in
problem.yaml. systest0 used this instead of the historic .timelimit
file, which caused some problems to have no time limit on the CDS until
we connected to the CCS, which caused problems for analytics.

This just adds support for reading the time limit from the yaml.