Fix properties when using a branding presentation
Presentations 'inside' a branding presentation were not receiving their
property changes, e.g. you couldn't set the text on a message presentation.
This is because each property is passed to a specific presentation, and
since e.g. branding pres != message pres, it wouldn't be told about the
property change in order to pass it on.

This makes the branding presentation special and told about all property
changes, and it in turn decides if the presentation should be

Ok, I know this is getting a little ugly, but still better to do the minimal
change for supporting 'branding' presentations until we know how we would
want to use them long term.