Fix issue resetting contest start
The contest info object is now in the default configuration so that we can
support extended properties like the rgb value. However, that means when a
CCS doesn't send a contest time (e.g. during a pause), the CDS will set the
default (original) value back. This puts in a temporary fix to never set
the default value for contest start time. Later we should revisit; using
the contest /access endpoint might be a better solution.