Increase segments in balloon map pres
When a team is far away from the contest site the arc a balloon takes to
travel there is noticeably a set of line segments in an arc approximating
a curve. We could get fancy and have a variable number of segments depending
on the distance or switch to an actual curve... but just increasing the
number of segments should be fine.