Default contest id from yaml
If a contest is configured in a folder where the contest id doesn't match
the folder name, and there is no CCS configured, and the contest id isn't
set in cdsConfig.xml, we should be picking up the default contest id from
the contest.yaml, not the folder name.

Adding support for this only for contest.yaml, not contest.json for now.
We could wait until the contest is loaded and let the disk contest source
load the id, but this delays too much, e.g. clients might be able to see
the endpoints or contest exist in the web pages first.