Use different names for event feed cache
Allows clients to provide a name when creating a REST connector. Having a
unique name means that the cache will be created in a different file,
avoiding two users loading the same cache. If no name is provided, the
user name is used since most users have different access roles.

This allows allows moving the CDS's use of different caches out of the
REST connector, since it shouldn't really be in there.