Change PENDING_COLOR to (66,114,245)
Change the PENDING_COLOR to a shade of blue instead of yellow
in order to improve accessibility for colorblind folks. The
current shade of yellow is very hard to distinguish from the
green SOLVED_COLOR.
This change is discussed in Issue #978. This color choice is
Option 2 in that discussion. Screenshots can be found there.
If you have further comments on the color choice, please
continue the discussion in Issue #978 rather than on this PR.