Autocorrect presentation admin CDS root with path
Since the presentation admin is 'global' and can control multiple contests
/presentations at once, the correct path to access the CDS is the CDS root,
e.g. something like http://cds/ or https://10.3.3.207:8443. People
frequently include the contest api in the url though, e.g.
https://cds/api/contests.

Although giving an error is 'correct', we've had a few requests to just
make it work. So instead of an error, this output a warning and tries the
url without the path.