global default properties - pres rows
The resolver docs needed to be updated to remove the --presenter and --client
options. Instead, I added information on what account types must be used, and
I thought it makes more sense (and is easier to explain) if we group the
resolver command line help by resolving options vs display options.