Always draw recent background in helper function
The helper method paintRecentBackground() was re-checking ContestUtil.isRecent(),
but the contest used might be updated concurrently, and in that case a white background was drawn
with a null Color, causing white flashing. The helper function is only used to draw recent updates,
so this renames the function to paintRecentResultBackground, and removes the extra check.