Do not show scoreboard before showing group winners photo screen
Initially we wanted to not show the scoreboard between two 'list' awards. However, this is not true in general. Imagine we want to show a list with all the silver medalists after resolving fifth place, but also a list of all gold medalists just after that. Then we DO want to show the scoreboard in between, since we want to show the highlighted teams. This made me decide to simplify things a lot: add a parameter to indicate whether we show the scoreboard in between. For now, we know this always comes AFTER another list, so we know we do not have to pause in between. I did add a note so we know what to change if that assumption ever changes.