Add back support for team photos to image generator
When the CMS started exporting contest archives I had to rewrite the prior logic
that pulled from the CMS file attachment format, and the last piece that I hadn't
updated was team photos. This removes the old CMS-folder-structure team photo
logic and reuses the same contest format function we do for logos and person
photos.

While I was doing that I fixed a few other things:
- Stop warning when team or person photos don't have transparency (code was there
  for logos).
- Fix out of date class header.
- Remove the MAX_IMAGE_SIZE image generation. It was there for a couple cases
  where we've had logos that were >4000px, but it is now already handled by
  generating a 1080p (desktop) logo.