Switch from UPC to QR in balloon printout
Switch from custom UPCa drawing code to using a QR code library. QR code
scanners aren't significantly more expensive than barcode scanners now,
it uses less space (so will conflict with messages less on the printout),
has built-in error correction, and we can use 'icpc-balloon-X' instead
of just a number.

I've used the qrcodegen library. The 1.8 release is a little stale, but the
github repo still has activity and it's not like QR codes have changed a
lot lately (and my UPC code was older). I'm putting the library in the
model project instead of balloon util so that it can be reused by other
tools (like snakeyaml or sentry).