Jakarta rebase
I haven't updated WLP in 2 years, and even prior to that Java EE changed
to Jakarta EE. The prior commit does the bulk renaming of packages and
classpath updates, this one deletes one missed jar, updates to Jakarta
features in server.xml, fixes the build, Dockerfile, web.xml, etc.

There could easily be some additional changes or something I've missed,
but this is the bulk of the change, works locally, and at the point I
need to see build failures or build artifacts to see if there's anything
else.