handle draft vs current spec
On the main branch the CDS has moved up to the draft spec, but for NAC and Baku
we will be running on the 2023_06 spec and there's no target to finalize the
current draft. In order to safely run these contests and release the CDS in the
interim, the default should be the 2023_06 spec.

This PR:

- Switches the default spec back to 2023_06.
- Changes the system property to 'ICPC_CONTEST_API=draft' in order to use the
  new/draft spec - the idea being that we always support the current released
  spec with an option to enable the new draft.
- Adds support to the /access endpoint to return the correct spec version.
- Adds a 'contest-api' property to the cdsConfig to make it easy to switch
  between specs, e.g. `<cds contest-api="draft" ...`