Signal if file operation fails
Alternative is to create a custom class inheriting from File and
overriding this function with the expected error handling.

This should never happen, but if this fails the user most likely wants
to know their storage is failing.

Kept the if without accolades as it seems to be the code style.