Delay startup of p100 to wait for the contestFeed reading
When starting P100 we expect the CDS to have a contest and P100 to get that contest in the feed, currently we would get the teamnumber, get the team in the Contest concept for the logo which would fail as the feed is not done reading yet.

The proper fix would be to get the "global" property to get if we're done reading the feed and only then starting the presentation (for some of them) but finding that thread was hard.