Request wider logos in balloon printout
Organization logos are supposed to be 'square-ish', but when they are more horizontal
they don't look great in the balloon printouts because we don't give them any more
space. We don't want to give these logos too much additional real-estate or squeeze
the title either, so this just bumps the width to 2x height.

Fixes #1124.